requirements.txt
opencv-python
torch
torchvision
transformers
ultralytics
pyttsx3
numpy
Pillow
